change name to caser
The binary doesn't necessarily need to be tied to kak, it is a useful
tool on it's own. I propose changing the name to 'caser' since it is
easier to remember and type, and it might help others find the tool
more easily.

change name to caser

The binary doesn't necessarily need to be tied to kak, it is a useful
tool on it's own. I propose changing the name to 'caser' since it is
easier to remember and type, and it might help others find the tool
more easily.